address moves
This commit was sponsored by Sergio Bost, Matt Campbell, and my other
patrons.  If you want to join them, you can support my work at
https://glyph.im/patrons/.
